requirements.txt
# requirements.txt - Add these lines
google-generativeai==0.3.2
newspaper3k==0.2.8
lxml[html]==4.9.3

# Keep existing requirements
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
